Support directory added. Moved simple scripts. Utils's class import modified
